Remove whitespace highlighting overrides in markdown file plugin.
The attempt to clear the augroup from the vimrc file works, but only after BufRead has already taken effect, so I end up with a highlight every time I type a space because InsertEnter and InsertLeave are cleared.

Just delete the whole thing, including the attempt to use trailing characters to highlight trailing spaces. The yellow highlights are tolerable on the rare times when I want to force a break with two spaces.